解决bug，新增ws异步通知结束

+++ go-server/src/main/java/com/goplatform/server/pojo/domain/KataCount.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class KataCount {
-    private int black;
+    private double black;
 
-    private int white;
+    private double white;
 }
